gui/controller/requirements.txt
Django==1.9.8
djangorestframework==3.4.0
django-cors-headers==1.1.0
django-jet

